requirements.txt
slack-bolt
slack-sdk
python-dotenv
-e libs/memex-core
